requirements.txt
# Anti-detection
curl_cffi>=0.7.0
camoufox[geoip]>=0.4.0
browserforge>=1.1.0

# Browser automation
playwright>=1.44.0

# Scraping / parsing
beautifulsoup4>=4.12.0
markdownify>=0.12.0
dateparser>=1.2.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.25.0

# AI
anthropic>=0.28.0
pydantic>=2.7.0
tenacity>=8.3.0

# Dynamic CV
WeasyPrint>=62.0
Jinja2>=3.1.0
python-docx>=1.1.0

# Storage
gspread>=6.1.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.130.0

# Utilities
python-dotenv>=1.0.0
filelock>=3.14.0
APScheduler>=3.10.0
numpy>=1.26.0
rich>=13.7.0
pathlib2>=2.3.0
PyYAML>=6.0.0
